pyproject.toml
[tool.poetry]
name = "brackend"
version = "0.1.0"
description = ""
authors = ["Peter Robicheaux <probicheaux@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.26.0"
Flask = "^2.0.2"
gunicorn = "^20.1.0"
Flask-Cors = "^3.0.10"
SQLAlchemy = "^1.4.27"
psycopg2 = "^2.9.2"
dramatiq = {extras = ["redis", "watch"], version = "^1.12.0"}
python-dotenv = "^0.19.2"
PyJWT = "^2.3.0"
firebase-admin = "^5.1.0"
bcrypt = "^3.2.0"
yagmail = "^0.14.260"

[tool.poetry.dev-dependencies]
black = "^21.9b0"
pylint = "^2.11.1"
docformatter = "^1.4"
jedi-language-server = "^0.34.6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
